Refactor Entropy map and take pathfinding distance into account when placing bases.
Also, now only use 0x11 for water tile textures, so that water shows up on the
minimap. Place a crater under all oil resources. Add some comments.

+++ src/map.cpp
@@ -784,7 +784,7 @@ bool mapLoad(char const *filename, bool preview)
 	ASSERT(psMapTiles == nullptr, "Map has not been cleared before calling mapLoad()!");
 
 	/* Allocate the memory for the map */
-	psMapTiles = (MAPTILE *)calloc(width * height, sizeof(MAPTILE));
+	psMapTiles = (MAPTILE *)calloc((size_t)width * height, sizeof(MAPTILE));
 	ASSERT(psMapTiles != nullptr, "Out of memory");
 
 	mapWidth = width;
@@ -882,7 +882,7 @@ bool mapLoadFromScriptData(ScriptMapData const &data, bool preview)
 	ASSERT(psMapTiles == nullptr, "Map has not been cleared before calling mapLoadFromScriptData()!");
 
 	/* Allocate the memory for the map */
-	psMapTiles = (MAPTILE *)calloc(data.mapWidth * data.mapHeight, sizeof(MAPTILE));
+	psMapTiles = (MAPTILE *)calloc((size_t)data.mapWidth * data.mapHeight, sizeof(MAPTILE));
 	ASSERT(psMapTiles != nullptr, "Out of memory");
 
 	mapWidth = data.mapWidth;
